split cv2.cpp: util, numpy
