+++ modules/python/src2/cv2.cpp
@@ -1,69 +1,18 @@
-//warning number '5033' not a valid compiler warning in vc12
-#if defined(_MSC_VER) && (_MSC_VER > 1800)
-// eliminating duplicated round() declaration
-#define HAVE_ROUND 1
-#pragma warning(push)
-#pragma warning(disable:5033)  // 'register' is no longer a supported storage class
-#endif
-
-// #define CVPY_DYNAMIC_INIT
-// #define Py_DEBUG
-
-#if defined(CVPY_DYNAMIC_INIT) && !defined(Py_DEBUG)
-#   define Py_LIMITED_API 0x03030000
-#endif
-
-#include <cmath>
-#include <Python.h>
-#include <limits>
-
-#if PY_MAJOR_VERSION < 3
-#undef CVPY_DYNAMIC_INIT
-#else
-#define CV_PYTHON_3 1
-#endif
-
-#if defined(_MSC_VER) && (_MSC_VER > 1800)
-#pragma warning(pop)
-#endif
-
-#define MODULESTR "cv2"
-#define NPY_NO_DEPRECATED_API NPY_1_7_API_VERSION
-
-#include <numpy/ndarrayobject.h>
+#include "cv2.hpp"
 
 #include "opencv2/opencv_modules.hpp"
 #include "opencv2/core.hpp"
-#include "opencv2/core/utils/configuration.private.hpp"
 #include "opencv2/core/utils/logger.hpp"
-#include "opencv2/core/utils/tls.hpp"
 
 #include "pyopencv_generated_include.h"
 #include "opencv2/core/types_c.h"
-#include "pycompat.hpp"
 #include <map>
 
 #include <type_traits>  // std::enable_if
 
 #define CV_HAS_CONVERSION_ERROR(x) (((x) == -1) && PyErr_Occurred())
 
-static PyObject* opencv_error = NULL;
-
 static PyTypeObject* pyopencv_Mat_TypePtr = nullptr;
-
-class ArgInfo
-{
-public:
-    const char* name;
-    bool outputarg;
-    // more fields may be added if necessary
-
-    ArgInfo(const char* name_, bool outputarg_) : name(name_), outputarg(outputarg_) {}
-
-private:
-    ArgInfo(const ArgInfo&) = delete;
-    ArgInfo& operator=(const ArgInfo&) = delete;
-};
 
 template<typename T, class TEnable = void>  // TEnable is used for SFINAE checks
 struct PyOpenCV_Converter
@@ -71,6 +20,8 @@ struct PyOpenCV_Converter
     //static inline bool to(PyObject* obj, T& p, const ArgInfo& info);
     //static inline PyObject* from(const T& src);
 };
+
+#include "cv2_util.hpp"
 
 // exception-safe pyopencv_to
 template<typename _Tp> static
@@ -98,405 +49,9 @@ bool pyopencv_to(PyObject* obj, T& p, const ArgInfo& info) { return PyOpenCV_Con
 template<typename T> static
 PyObject* pyopencv_from(const T& src) { return PyOpenCV_Converter<T>::from(src); }
 
-static bool isPythonBindingsDebugEnabled()
-{
-    static bool param_debug = cv::utils::getConfigurationParameterBool("OPENCV_PYTHON_DEBUG", false);
-    return param_debug;
-}
 
-static void emit_failmsg(PyObject * exc, const char *msg)
-{
-    static bool param_debug = isPythonBindingsDebugEnabled();
-    if (param_debug)
-    {
-        CV_LOG_WARNING(NULL, "Bindings conversion failed: " << msg);
-    }
-    PyErr_SetString(exc, msg);
-}
-
-static int failmsg(const char *fmt, ...)
-{
-    char str[1000];
-
-    va_list ap;
-    va_start(ap, fmt);
-    vsnprintf(str, sizeof(str), fmt, ap);
-    va_end(ap);
-
-    emit_failmsg(PyExc_TypeError, str);
-    return 0;
-}
-
-static PyObject* failmsgp(const char *fmt, ...)
-{
-    char str[1000];
-
-    va_list ap;
-    va_start(ap, fmt);
-    vsnprintf(str, sizeof(str), fmt, ap);
-    va_end(ap);
-
-    emit_failmsg(PyExc_TypeError, str);
-    return 0;
-}
-
-class PyAllowThreads
-{
-public:
-    PyAllowThreads() : _state(PyEval_SaveThread()) {}
-    ~PyAllowThreads()
-    {
-        PyEval_RestoreThread(_state);
-    }
-private:
-    PyThreadState* _state;
-};
-
-class PyEnsureGIL
-{
-public:
-    PyEnsureGIL() : _state(PyGILState_Ensure()) {}
-    ~PyEnsureGIL()
-    {
-        PyGILState_Release(_state);
-    }
-private:
-    PyGILState_STATE _state;
-};
-
-/**
- * Light weight RAII wrapper for `PyObject*` owning references.
- * In comparisson to C++11 `std::unique_ptr` with custom deleter, it provides
- * implicit conversion functions that might be useful to initialize it with
- * Python functions those returns owning references through the `PyObject**`
- * e.g. `PyErr_Fetch` or directly pass it to functions those want to borrow
- * reference to object (doesn't extend object lifetime) e.g. `PyObject_Str`.
- */
-class PySafeObject
-{
-public:
-    PySafeObject() : obj_(NULL) {}
-
-    explicit PySafeObject(PyObject* obj) : obj_(obj) {}
-
-    ~PySafeObject()
-    {
-        Py_CLEAR(obj_);
-    }
-
-    operator PyObject*()
-    {
-        return obj_;
-    }
-
-    operator PyObject**()
-    {
-        return &obj_;
-    }
-
-    PyObject* release()
-    {
-        PyObject* obj = obj_;
-        obj_ = NULL;
-        return obj;
-    }
-
-private:
-    PyObject* obj_;
-
-    // Explicitly disable copy operations
-    PySafeObject(const PySafeObject*); // = delete
-    PySafeObject& operator=(const PySafeObject&); // = delete
-};
-
-static void pyRaiseCVException(const cv::Exception &e)
-{
-    PyObject_SetAttrString(opencv_error, "file", PyString_FromString(e.file.c_str()));
-    PyObject_SetAttrString(opencv_error, "func", PyString_FromString(e.func.c_str()));
-    PyObject_SetAttrString(opencv_error, "line", PyInt_FromLong(e.line));
-    PyObject_SetAttrString(opencv_error, "code", PyInt_FromLong(e.code));
-    PyObject_SetAttrString(opencv_error, "msg", PyString_FromString(e.msg.c_str()));
-    PyObject_SetAttrString(opencv_error, "err", PyString_FromString(e.err.c_str()));
-    PyErr_SetString(opencv_error, e.what());
-}
-
-#define ERRWRAP2(expr) \
-try \
-{ \
-    PyAllowThreads allowThreads; \
-    expr; \
-} \
-catch (const cv::Exception &e) \
-{ \
-    pyRaiseCVException(e); \
-    return 0; \
-} \
-catch (const std::exception &e) \
-{ \
-    PyErr_SetString(opencv_error, e.what()); \
-    return 0; \
-} \
-catch (...) \
-{ \
-    PyErr_SetString(opencv_error, "Unknown C++ exception from OpenCV code"); \
-    return 0; \
-}
 
 using namespace cv;
-
-
-namespace {
-template<class T>
-NPY_TYPES asNumpyType()
-{
-    return NPY_OBJECT;
-}
-
-template<>
-NPY_TYPES asNumpyType<bool>()
-{
-    return NPY_BOOL;
-}
-
-#define CV_GENERATE_INTEGRAL_TYPE_NPY_CONVERSION(src, dst) \
-    template<>                                             \
-    NPY_TYPES asNumpyType<src>()                           \
-    {                                                      \
-        return NPY_##dst;                                  \
-    }                                                      \
-    template<>                                             \
-    NPY_TYPES asNumpyType<u##src>()                        \
-    {                                                      \
-        return NPY_U##dst;                                 \
-    }
-
-CV_GENERATE_INTEGRAL_TYPE_NPY_CONVERSION(int8_t, INT8);
-
-CV_GENERATE_INTEGRAL_TYPE_NPY_CONVERSION(int16_t, INT16);
-
-CV_GENERATE_INTEGRAL_TYPE_NPY_CONVERSION(int32_t, INT32);
-
-CV_GENERATE_INTEGRAL_TYPE_NPY_CONVERSION(int64_t, INT64);
-
-#undef CV_GENERATE_INTEGRAL_TYPE_NPY_CONVERSION
-
-template<>
-NPY_TYPES asNumpyType<float>()
-{
-    return NPY_FLOAT;
-}
-
-template<>
-NPY_TYPES asNumpyType<double>()
-{
-    return NPY_DOUBLE;
-}
-
-template <class T>
-PyArray_Descr* getNumpyTypeDescriptor()
-{
-    return PyArray_DescrFromType(asNumpyType<T>());
-}
-
-template <>
-PyArray_Descr* getNumpyTypeDescriptor<size_t>()
-{
-#if SIZE_MAX == ULONG_MAX
-    return PyArray_DescrFromType(NPY_ULONG);
-#elif SIZE_MAX == ULLONG_MAX
-    return PyArray_DescrFromType(NPY_ULONGLONG);
-#else
-    return PyArray_DescrFromType(NPY_UINT);
-#endif
-}
-
-template <class T, class U>
-bool isRepresentable(U value) {
-    return (std::numeric_limits<T>::min() <= value) && (value <= std::numeric_limits<T>::max());
-}
-
-template<class T>
-bool canBeSafelyCasted(PyObject* obj, PyArray_Descr* to)
-{
-    return PyArray_CanCastTo(PyArray_DescrFromScalar(obj), to) != 0;
-}
-
-
-template<>
-bool canBeSafelyCasted<size_t>(PyObject* obj, PyArray_Descr* to)
-{
-    PyArray_Descr* from = PyArray_DescrFromScalar(obj);
-    if (PyArray_CanCastTo(from, to))
-    {
-        return true;
-    }
-    else
-    {
-        // False negative scenarios:
-        // - Signed input is positive so it can be safely cast to unsigned output
-        // - Input has wider limits but value is representable within output limits
-        // - All the above
-        if (PyDataType_ISSIGNED(from))
-        {
-            int64_t input = 0;
-            PyArray_CastScalarToCtype(obj, &input, getNumpyTypeDescriptor<int64_t>());
-            return (input >= 0) && isRepresentable<size_t>(static_cast<uint64_t>(input));
-        }
-        else
-        {
-            uint64_t input = 0;
-            PyArray_CastScalarToCtype(obj, &input, getNumpyTypeDescriptor<uint64_t>());
-            return isRepresentable<size_t>(input);
-        }
-        return false;
-    }
-}
-
-
-template<class T>
-bool parseNumpyScalar(PyObject* obj, T& value)
-{
-    if (PyArray_CheckScalar(obj))
-    {
-        // According to the numpy documentation:
-        // There are 21 statically-defined PyArray_Descr objects for the built-in data-types
-        // So descriptor pointer is not owning.
-        PyArray_Descr* to = getNumpyTypeDescriptor<T>();
-        if (canBeSafelyCasted<T>(obj, to))
-        {
-            PyArray_CastScalarToCtype(obj, &value, to);
-            return true;
-        }
-    }
-    return false;
-}
-
-TLSData<std::vector<std::string> > conversionErrorsTLS;
-
-inline void pyPrepareArgumentConversionErrorsStorage(std::size_t size)
-{
-    std::vector<std::string>& conversionErrors = conversionErrorsTLS.getRef();
-    conversionErrors.clear();
-    conversionErrors.reserve(size);
-}
-
-void pyRaiseCVOverloadException(const std::string& functionName)
-{
-    const std::vector<std::string>& conversionErrors = conversionErrorsTLS.getRef();
-    const std::size_t conversionErrorsCount = conversionErrors.size();
-    if (conversionErrorsCount > 0)
-    {
-        // In modern std libraries small string optimization is used = no dynamic memory allocations,
-        // but it can be applied only for string with length < 18 symbols (in GCC)
-        const std::string bullet = "\n - ";
-
-        // Estimate required buffer size - save dynamic memory allocations = faster
-        std::size_t requiredBufferSize = bullet.size() * conversionErrorsCount;
-        for (std::size_t i = 0; i < conversionErrorsCount; ++i)
-        {
-            requiredBufferSize += conversionErrors[i].size();
-        }
-
-        // Only string concatenation is required so std::string is way faster than
-        // std::ostringstream
-        std::string errorMessage("Overload resolution failed:");
-        errorMessage.reserve(errorMessage.size() + requiredBufferSize);
-        for (std::size_t i = 0; i < conversionErrorsCount; ++i)
-        {
-            errorMessage += bullet;
-            errorMessage += conversionErrors[i];
-        }
-        cv::Exception exception(CV_StsBadArg, errorMessage, functionName, "", -1);
-        pyRaiseCVException(exception);
-    }
-    else
-    {
-        cv::Exception exception(CV_StsInternal, "Overload resolution failed, but no errors reported",
-                                functionName, "", -1);
-        pyRaiseCVException(exception);
-    }
-}
-
-void pyPopulateArgumentConversionErrors()
-{
-    if (PyErr_Occurred())
-    {
-        PySafeObject exception_type;
-        PySafeObject exception_value;
-        PySafeObject exception_traceback;
-        PyErr_Fetch(exception_type, exception_value, exception_traceback);
-        PyErr_NormalizeException(exception_type, exception_value,
-                                 exception_traceback);
-
-        PySafeObject exception_message(PyObject_Str(exception_value));
-        std::string message;
-        getUnicodeString(exception_message, message);
-#ifdef CV_CXX11
-        conversionErrorsTLS.getRef().push_back(std::move(message));
-#else
-        conversionErrorsTLS.getRef().push_back(message);
-#endif
-    }
-}
-
-struct SafeSeqItem
-{
-    PyObject * item;
-    SafeSeqItem(PyObject *obj, size_t idx) { item = PySequence_GetItem(obj, idx); }
-    ~SafeSeqItem() { Py_XDECREF(item); }
-
-private:
-    SafeSeqItem(const SafeSeqItem&); // = delete
-    SafeSeqItem& operator=(const SafeSeqItem&); // = delete
-};
-
-template <class T>
-class RefWrapper
-{
-public:
-    RefWrapper(T& item) : item_(item) {}
-
-    T& get() CV_NOEXCEPT { return item_; }
-
-private:
-    T& item_;
-};
-
-// In order to support this conversion on 3.x branch - use custom reference_wrapper
-// and C-style array instead of std::array<T, N>
-template <class T, std::size_t N>
-bool parseSequence(PyObject* obj, RefWrapper<T> (&value)[N], const ArgInfo& info)
-{
-    if (!obj || obj == Py_None)
-    {
-        return true;
-    }
-    if (!PySequence_Check(obj))
-    {
-        failmsg("Can't parse '%s'. Input argument doesn't provide sequence "
-                "protocol", info.name);
-        return false;
-    }
-    const std::size_t sequenceSize = PySequence_Size(obj);
-    if (sequenceSize != N)
-    {
-        failmsg("Can't parse '%s'. Expected sequence length %lu, got %lu",
-                info.name, N, sequenceSize);
-        return false;
-    }
-    for (std::size_t i = 0; i < N; ++i)
-    {
-        SafeSeqItem seqItem(obj, i);
-        if (!pyopencv_to(seqItem.item, value[i].get(), info))
-        {
-            failmsg("Can't parse '%s'. Sequence item with index %lu has a "
-                    "wrong type", info.name, i);
-            return false;
-        }
-    }
-    return true;
-}
-} // namespace
 
 namespace traits {
 template <bool Value>
@@ -559,78 +114,7 @@ typedef std::vector<std::vector<Point3f> > vector_vector_Point3f;
 typedef std::vector<std::vector<DMatch> > vector_vector_DMatch;
 typedef std::vector<std::vector<KeyPoint> > vector_vector_KeyPoint;
 
-class NumpyAllocator : public MatAllocator
-{
-public:
-    NumpyAllocator() { stdAllocator = Mat::getStdAllocator(); }
-    ~NumpyAllocator() {}
-
-    UMatData* allocate(PyObject* o, int dims, const int* sizes, int type, size_t* step) const
-    {
-        UMatData* u = new UMatData(this);
-        u->data = u->origdata = (uchar*)PyArray_DATA((PyArrayObject*) o);
-        npy_intp* _strides = PyArray_STRIDES((PyArrayObject*) o);
-        for( int i = 0; i < dims - 1; i++ )
-            step[i] = (size_t)_strides[i];
-        step[dims-1] = CV_ELEM_SIZE(type);
-        u->size = sizes[0]*step[0];
-        u->userdata = o;
-        return u;
-    }
-
-    UMatData* allocate(int dims0, const int* sizes, int type, void* data, size_t* step, AccessFlag flags, UMatUsageFlags usageFlags) const CV_OVERRIDE
-    {
-        if( data != 0 )
-        {
-            // issue #6969: CV_Error(Error::StsAssert, "The data should normally be NULL!");
-            // probably this is safe to do in such extreme case
-            return stdAllocator->allocate(dims0, sizes, type, data, step, flags, usageFlags);
-        }
-        PyEnsureGIL gil;
-
-        int depth = CV_MAT_DEPTH(type);
-        int cn = CV_MAT_CN(type);
-        const int f = (int)(sizeof(size_t)/8);
-        int typenum = depth == CV_8U ? NPY_UBYTE : depth == CV_8S ? NPY_BYTE :
-        depth == CV_16U ? NPY_USHORT : depth == CV_16S ? NPY_SHORT :
-        depth == CV_32S ? NPY_INT : depth == CV_32F ? NPY_FLOAT :
-        depth == CV_64F ? NPY_DOUBLE : f*NPY_ULONGLONG + (f^1)*NPY_UINT;
-        int i, dims = dims0;
-        cv::AutoBuffer<npy_intp> _sizes(dims + 1);
-        for( i = 0; i < dims; i++ )
-            _sizes[i] = sizes[i];
-        if( cn > 1 )
-            _sizes[dims++] = cn;
-        PyObject* o = PyArray_SimpleNew(dims, _sizes.data(), typenum);
-        if(!o)
-            CV_Error_(Error::StsError, ("The numpy array of typenum=%d, ndims=%d can not be created", typenum, dims));
-        return allocate(o, dims0, sizes, type, step);
-    }
-
-    bool allocate(UMatData* u, AccessFlag accessFlags, UMatUsageFlags usageFlags) const CV_OVERRIDE
-    {
-        return stdAllocator->allocate(u, accessFlags, usageFlags);
-    }
-
-    void deallocate(UMatData* u) const CV_OVERRIDE
-    {
-        if(!u)
-            return;
-        PyEnsureGIL gil;
-        CV_Assert(u->urefcount >= 0);
-        CV_Assert(u->refcount >= 0);
-        if(u->refcount == 0)
-        {
-            PyObject* o = (PyObject*)u->userdata;
-            Py_XDECREF(o);
-            delete u;
-        }
-    }
-
-    const MatAllocator* stdAllocator;
-};
-
-NumpyAllocator g_numpyAllocator;
+#include "cv2_numpy.hpp"
 
 
 enum { ARG_NONE = 0, ARG_MAT = 1, ARG_SCALAR = 2 };
